Add loggers features to mc-common
