+++ common/Cargo.toml
@@ -14,18 +14,21 @@ std = [
 log = [
     "backtrace",
     "chrono",
+    "mc-util-logger-macros",
+    "slog",
+    "slog-scope",
+]
+loggers = [
+    "log",
     "hostname",
     "lazy_static",
     "mc-util-build-info",
-    "mc-util-logger-macros",
     "sentry",
-    "slog",
     "slog-async",
     "slog-atomic",
     "slog-envlogger",
     "slog-gelf",
     "slog-json",
-    "slog-scope",
     "slog-stdlog",
     "slog-term",
 ]
@@ -48,21 +51,23 @@ serde = { version = "1.0", default-features = false, features = ["alloc", "deriv
 sha3 = { version = "0.8", default-features = false }
 siphasher = "0.3"
 
-# log-only dependencies
+# log- and loggers-only dependencies
 backtrace = { version = "0.3", optional = true }
-mc-util-build-info = { path = "../util/build/info", optional = true }
 chrono = { version = "0.4", optional = true }
+mc-util-logger-macros = { path = "../util/logger-macros", optional = true }
+slog = { version = "2.5", features = ["dynamic-keys", "max_level_trace", "release_max_level_trace"], optional = true }
+slog-scope = { version = "4.1.2", optional = true }
+
+# loggers-only dependencies
 hostname = { version = "0.1", optional = true }
 lazy_static = { version = "1.4", optional = true }
-mc-util-logger-macros = { path = "../util/logger-macros", optional = true }
+mc-util-build-info = { path = "../util/build/info", optional = true }
 sentry = { version = "0.18", optional = true, default-features = false, features = ["with_client_implementation", "with_reqwest_transport", "with_panic", "with_failure", "with_device_info", "with_rust_info", "with_rustls"] }
-slog = { version = "2.5", features = ["dynamic-keys", "max_level_trace", "release_max_level_trace"], optional = true }
 slog-async = { version = "2.3", optional = true }
 slog-atomic = { version = "3.0", optional = true }
 slog-envlogger = { version = "2.1", optional = true }
 slog-gelf = { version = "0.1", optional = true }
 slog-json = { version = "2.3", optional = true }
-slog-scope = { version = "4.1.2", optional = true }
 slog-stdlog = { version = "4.0.0", optional = true }
 slog-term = { version = "2.4", optional = true }
